### 2019-06-06 #1489 @labarba
👋 @AliYoussef96 — You need to put your references into a `.bib` file, and add this to the metadata of the paper, as shown in the [Author Guide](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography).

Also, on a first look at your paper, I want to point out that Summary section throws the reader into the deep end with technical background and lots of jargon. Please give this another pass: we would like the paper to give a reader an idea of what the software does and what research context it is used for. Take note of this advice from the [Author Guide](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain):

> JOSS welcomes submissions from broadly diverse research areas. For this reason, we require that authors include in the paper some sentences that would explain the software functionality and domain of use to a non-specialist reader.
