### 2019-06-06 #1489 @labarba
Folks — the author suggested @csoneson as editor, but given her focus on R (this is a Python package), maybe another editor can best handle this? Could either @lpantano or @mgymrek take this submission? Let me know!
