### 2019-06-06 e6f82c2
Update paper.md
