+++ paper.md
@@ -12,6 +12,7 @@ affiliations:
  - name: Department of Genetics, Faculty of Agriculture, Cairo University, 12613, Cairo, Egypt
    index: 1
 date: June 3, 2019
+bibliography: paper.bib
 ---
 # Summary
 
