### 2019-06-06 c3eace7
Create paper.bib
