+++ paper.bib
@@ -1,0 +1,37 @@
+@article{Sharp1987,
+  doi = {10.1093/nar/15.3.1281},
+  url = {https://doi.org/10.1093/nar/15.3.1281},
+  year  = {1987},
+  publisher = {Oxford University Press ({OUP})},
+  volume = {15},
+  number = {3},
+  pages = {1281--1295},
+  author = {Paul M. Sharp and Wen-Hsiung Li},
+  title = {The codon adaptation index-a measure of directional synonymous codon usage bias, and its potential applications},
+  journal = {Nucleic Acids Research}
+}
+
+@article{Lee. 2018,
+  doi = {10.21105/joss.00905},
+  url = {https://doi.org/10.21105/joss.00905},
+  year  = {2018},
+  publisher = {Journal of Open Source Software ({JOSS})},
+  volume = {3},
+  number = {30},
+  author = {Benjamin D. Lee},
+  title = {Python Implementation of Codon Adaptation Index},
+  journal = {Journal of Open Source Software}
+}
+
+@article{Wright1990,
+  doi = {10.1016/0378-1119(90)90491-9},
+  url = {https://doi.org/10.1016/0378-1119(90)90491-9},
+  year  = {1990},
+  publisher = {Elsevier B.V},
+  volume = {87},
+  number = {1},
+  pages = {23--29},
+  author = {Frank Wright},
+  title = {The ‘effective number of codons’ used in a gene},
+  journal = {Gene}
+}
