+++ paper.bib
@@ -1,4 +1,4 @@
-@article{Sharp1987,
+@article{Sharp and Li 1987,
   doi = {10.1093/nar/15.3.1281},
   url = {https://doi.org/10.1093/nar/15.3.1281},
   year  = {1987},
@@ -23,7 +23,7 @@
   journal = {Journal of Open Source Software}
 }
 
-@article{Wright1990,
+@article{Wright 1990,
   doi = {10.1016/0378-1119(90)90491-9},
   url = {https://doi.org/10.1016/0378-1119(90)90491-9},
   year  = {1990},
@@ -34,4 +34,30 @@
   author = {Frank Wright},
   title = {The ‘effective number of codons’ used in a gene},
   journal = {Gene}
+}
+
+@article{Wang et al., 2018,
+  doi = {10.1371/journal.pone.0194372},
+  url = {https://doi.org/10.1371/journal.pone.0194372},
+  year  = {2018},
+  publisher = {PLOS},
+  volume = {13},
+  number = {3},
+  pages = {1--17},
+  author = {Liyuan Wang , Huixian Xing , Yanchao Yuan, Xianlin Wang, Muhammad Saeed, Jincai Tao, Wei Feng, Guihua Zhang, Xianliang Song and Xuezhen Sun},
+  title = {Genome-wide analysis of codon usage bias in four sequenced cotton species},
+  journal = {PLoS ONE}
+}
+
+@article{Song et al. 2017,
+  doi = {10.3389/fmicb.2017.01419},
+  url = {https://doi.org/10.3389/fmicb.2017.01419},
+  year  = {2017},
+  publisher = {Frontiers},
+  volume = {8},
+  number = {6},
+  pages = {1--12},
+  author = {Hui Song, Jing Liu, Qiuyan Song, Qingping Zhang, Pei Tian and Zhibiao Nan},
+  title = {Comprehensive Analysis of Codon Usage Bias in Seven Epichloë Species and Their Peramine-Coding Genes},
+  journal = {Frontiers in Microbiology}
 }
