+++ paper.bib
@@ -61,3 +61,16 @@
   title = {Comprehensive Analysis of Codon Usage Bias in Seven Epichloë Species and Their Peramine-Coding Genes},
   journal = {Frontiers in Microbiology}
 }
+
+@article{Hershberg and Petrov 2009,
+  doi = {10.1371/journal.pgen.1000556},
+  url = {https://doi.org/10.1371/journal.pgen.1000556},
+  year  = {2009},
+  publisher = {PLOS},
+  volume = {5},
+  number = {7},
+  pages = {1--10},
+  author = {Hershberg, Ruth, and Dmitri A. Petrov},
+  title = {General Rules for Optimal Codon Choice},
+  journal = {PLoS Genetics}
+}
