### 2019-06-06 #1489 @labarba
Thanks for agreeing to review this JOSS submission, @Benjamin-Lee! I will assign you now, but hold your horses, as we haven't assigned the handling editor yet (and they'll want a second reviewer).

I'm waiting for either @lpantano or @mgymrek to agree to edit.
