+++ paper.bib
@@ -1,4 +1,4 @@
-@article{Sharp and Li 1987,
+@article{Sharp1987,
   doi = {10.1093/nar/15.3.1281},
   url = {https://doi.org/10.1093/nar/15.3.1281},
   year  = {1987},
@@ -11,7 +11,7 @@
   journal = {Nucleic Acids Research}
 }
 
-@article{Lee. 2018,
+@article{Lee2018,
   doi = {10.21105/joss.00905},
   url = {https://doi.org/10.21105/joss.00905},
   year  = {2018},
@@ -23,7 +23,7 @@
   journal = {Journal of Open Source Software}
 }
 
-@article{Wright 1990,
+@article{Wright1990,
   doi = {10.1016/0378-1119(90)90491-9},
   url = {https://doi.org/10.1016/0378-1119(90)90491-9},
   year  = {1990},
@@ -36,7 +36,7 @@
   journal = {Gene}
 }
 
-@article{Wang et al., 2018,
+@article{Wang2018,
   doi = {10.1371/journal.pone.0194372},
   url = {https://doi.org/10.1371/journal.pone.0194372},
   year  = {2018},
@@ -49,7 +49,7 @@
   journal = {PLoS ONE}
 }
 
-@article{Song et al. 2017,
+@article{Song2017,
   doi = {10.3389/fmicb.2017.01419},
   url = {https://doi.org/10.3389/fmicb.2017.01419},
   year  = {2017},
@@ -62,7 +62,7 @@
   journal = {Frontiers in Microbiology}
 }
 
-@article{Hershberg and Petrov 2009,
+@article{Hershberg2009,
   doi = {10.1371/journal.pgen.1000556},
   url = {https://doi.org/10.1371/journal.pgen.1000556},
   year  = {2009},
