### 2019-06-09 #1500 @whedon
[REVIEW]: BCAW: Automated tool for codon usage bias analysis for molecular evolution
**Submitting author:** @AliYoussef96 (<a href="http://orcid.org/0000-0002-5201-387X">Ali Youssef</a>)
**Repository:** <a href="https://github.com/AliYoussef96/BCAW-Tool" target ="_blank">https://github.com/AliYoussef96/BCAW-Tool</a>
**Version:** v1.0.3
