+++ README.md
@@ -6,7 +6,9 @@ Expected results from BCAW Tool are; nucleotide content, various codon usage mea
 BCAW Tool manages a complete automated workflow to analysis the codon usage bias for genes and genomes of any organism. BCAW Tool is available as executable application, work under Windows operating system, also a source code is available.
 ## Dependencies
 
-This software was designed by python 3.7 if it will be used as an executable file on Windows operating system no need for any dependencies. If the source code will be used, these packages must be installed:
+This software was designed by python 3.7 if it will be used as an executable application on Windows operating system no need for any dependencies.
+
+** If the source code will be used, these packages must be installed:
 
 1- Biopython
 
@@ -21,6 +23,7 @@ This software was designed by python 3.7 if it will be used as an executable fil
 6- numpy
 
 ## Installation Instructions
+** To run as an executable application:
 
 1- Download (BCAW Tool.zip) file available at: https://sourceforge.net/projects/bcaw-tool/
 
@@ -29,6 +32,14 @@ This software was designed by python 3.7 if it will be used as an executable fil
 3- Within the BCAW Tool folder you expect to find; BCAW tool folder, Source Code folder, Result folder shortcut, BCAW tool.exe shortcut, and pdf tutorial file.
 
 4- Run the tool from BCAW tool.exe shortcut icon.
+
+** To run as python application from the source code:
+
+1- Download the source code folder available in this repository ( You must expect to find 11 python files there ).
+
+2- Install the dependencies, mentioned above.
+
+3- Run the BCAWT.py file.
 
 ## Contribution Guidelines
 To report bugs or seek support please open an issue on this repository. Contributions to the software are welcome; please open an issue for further discussion.
