### 2019-06-26 #1500 @Benjamin-Lee
Ok, I just added five initial issues to the repo. @AliYoussef96, I hope my issues are understandable but if not, please don't hesitate to reach out for more information. I'm excited to try this tool out!
