### 2019-06-26 #1500 @mgymrek
Hi @AliYoussef96

I agree with the points raised by @Benjamin-Lee. See also my comments below.

### Install process
I agree this should be structured more like a standard Python library with a setup.py file for easy installation of the tool and dependencies.

### Usage/docs
1. I personally prefer that tools take in all the arguments at once rather than prompting for each one. That way it can be more easily embedded in other pipelines that might not directly interact with uses. You could also give a complete command on example data for someone to try.
2. It would be helpful if the README specified in the initial description what are the inputs and outputs. It looks like the input is a fasta file with gene sequences, rather than just any fasta file (e.g. with a reference genome). It could also be helpful to point users to how to obtain such fasta files for a species of interest.
3. I could not find documentation of what the plots are showing. e.g. what do GC1, GC2, GC3 mean for the first plot shown and what do Axis 1 vs. Axis 2 mean in the second plot?

### README and paper:
REAMDE: "automated workflow to analysis the codon usage " -> "automated workflow to analyze the codon usage'
There are now references in the paper. It would also be helpful for the paper to describe in more detail what the outputs are.

### Testing
I do not see tests. It would be great to both tests functions with automated tests, and also to provide some indication that your overall results are correct. e.g. by comparing E. Coli results to previously published ones.
