+++ README.md
@@ -1,14 +1,12 @@
 # BCAW: Automated tool for codon usage bias analysis for molecular evolution
 
 ## Statement of Need
+
 There are no tools available enable users to run a whole automated workflow for codon usage bias analysis. Using python 3.7 BCAW Tool ( Bio Codon Analysis Workflow Tool ) was developed to address this problem.
 Expected results from BCAW Tool are; nucleotide content, various codon usage measures (as Effective number of codons, Codon adaptation index, etc.) as well as correlation analysis with graphs. Further, Correspondence Analysis and determination of putative optimal codons are included.
 BCAW Tool manages a complete automated workflow to analysis the codon usage bias for genes and genomes of any organism. BCAW Tool is available as executable application, work under Windows operating system, also a source code is available.
+
 ## Dependencies
-
-This software was designed by python 3.7 if it will be used as an executable application on Windows operating system no need for any dependencies.
-
-**If the source code will be used, these packages must be installed:**
 
 1- Biopython
 
@@ -23,101 +21,42 @@ This software was designed by python 3.7 if it will be used as an executable app
 6- numpy
 
 ## Installation Instructions
-**To run as an executable application:**
 
-1- Download (BCAW Tool.zip) file available at: https://sourceforge.net/projects/bcaw-tool/
 
-2- Extract the files.
+**Using pip**
 
-3- Within the BCAW Tool folder you expect to find; BCAW tool folder, Source Code folder, Result folder shortcut, BCAW tool.exe shortcut, and pdf tutorial file.
-
-4- Run the tool from BCAW tool.exe shortcut icon.
-
-**To run as python application from the source code:**
-
-1- Download the source code folder available in this repository ( You must expect to find 11 python files there ).
-
-2- Install the dependencies, mentioned above.
-
-3- Run the BCAWT.py file.
+```
+pip install
+```
 
 ## Contribution Guidelines
 To report bugs or seek support please open an issue on this repository. Contributions to the software are welcome; please open an issue for further discussion.
 
-## How to Use BCAW tool?
-Unzip the file you downloaded, run the tool from BCAWT.exe shortcut icon then ; 
+Usage
+=====
 
-**The first message from the tool will be:**
-```
--s or -m then Enter the name of the file:
-```
-              
-Where user has to input the main ( fasta ) file, here there are two options  to be used;
+**Auto testing**
 
-The first one:
-```
--s FileName
-```             
-FileName refers to a path of ( fasta ) format file contains genes user want to be analyzed. 
-Note: Enter filename without ( .fasta )
-Example:
-```
--s D:\x\y\BCAW Tool\sample_input_files\Escherichia coli str. K-12 substr. MG1655
-```
-The second one:
-```
--m FileName
-```
-This time FileName refers to a path of ( text ) format file contains paths for ( fasta ) files containing genes to be analyzed for different organisms.
+First download fasta file containing the coding sequence
 
-Note: enter filename without ( .txt )
+[Test file](https://github.com/AliYoussef96/BCAWTool/blob/master/Ecoli.fasta)
 
-Example:
-```
--m D:\x\y\BCAW Tool\sample_input_files\ Genomes
-```
-**The second message from the tool will be:**
-```
--s or -m Enter the name of the reference gene set file or -Auto:
-```
-Where user has to input the reference genes set ( fasta ) file, here there are three options  to be used; 
+then run ( It will automatically run a test on the results files ):
 
-The first one:
 ```
--s FileName_reference
+from BCAWT import BCAWT_auto_test
+BCAWT_auto_test.auto_test()
+BCAWT_auto_test.auto_check_files()
[…]
