+++ docs/install.rst
@@ -1,0 +1,8 @@
+Installation
+===============
+
+*Using pip*
+
+	pip install
+
+or by downloading the source code from https://github.com/AliYoussef96/BCAW-Tool
