+++ docs/help.rst
@@ -1,0 +1,3 @@
+Help
+========
+If you need more help please contact ali.mo.anwar@std.agr.cu.edu.eg
