+++ docs/modules.rst
@@ -1,0 +1,7 @@
+BCAWT
+=====
+
+.. toctree::
+   :maxdepth: 4
+
+   BCAWT
