+++ README.md
@@ -57,10 +57,15 @@ from BCAWT import BCAWT
 BCAWT.BCAW('Ecoli','result_folder',fasta=True,Auto=True)
 >> Results Saved
 ```
+**Input**
 
-## Expected Output files
+input_the_main_fasta_file (str) – fasta file contains DNA sequence 
 
+input_the_ref_fasta_file (str) – fasta file contains reference DNA sequence, default = None
+
+**Output**
 ![Table 1: Expected Output files](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Table.png)
+
 
 ## Output examples
 
