+++ docs/conf.py
@@ -30,7 +30,7 @@ release = '0.1.0'
 # Add any Sphinx extension module names here, as strings. They can be
 # extensions coming with Sphinx (named 'sphinx.ext.*') or your custom
 # ones.
-extensions = ["sphinxcontrib.napoleon"
+extensions = [
 ]
 
 # Add any paths that contain templates here, relative to this directory.
