### 2019-06-27 f2f834a
Delete conf.py
