+++ docs/Usage.rst
@@ -9,15 +9,14 @@ https://github.com/AliYoussef96/BCAWTool/blob/master/Ecoli.fasta
 
 then run ( It will automatically run a test on the results files ):
 
+
 ```
 from BCAWT import BCAWT_auto_test
-    
 BCAWT_auto_test.auto_test()
-    
 BCAWT_auto_test.auto_check_files()
-    
 >> test is completed 'successfully'
 ```
+
 
 **Main Usage**
 
