+++ docs/Usage.rst
@@ -10,23 +10,17 @@ https://github.com/AliYoussef96/BCAWTool/blob/master/Ecoli.fasta
 then run ( It will automatically run a test on the results files ):
 
 
-```
-from BCAWT import BCAWT_auto_test
 
-BCAWT_auto_test.auto_test()
+  from BCAWT import BCAWT_auto_test
+  BCAWT_auto_test.auto_test()
+  BCAWT_auto_test.auto_check_files()
+  >> test is completed 'successfully'
 
-BCAWT_auto_test.auto_check_files()
-
->> test is completed 'successfully'
-```
 
 
 **Main Usage**
 
-```
-from BCAWT import BCAWT
+  from BCAWT import BCAWT
+  BCAWT.BCAW('Ecoli','result_folder',fasta=True,Auto=True)
+  >> Results Saved
 
-BCAWT.BCAW('Ecoli','result_folder',fasta=True,Auto=True)
-
->> Results Saved
-```
