+++ docs/Usage.rst
@@ -12,8 +12,11 @@ then run ( It will automatically run a test on the results files ):
 
 
   from BCAWT import BCAWT_auto_test
+  
   BCAWT_auto_test.auto_test()
+  
   BCAWT_auto_test.auto_check_files()
+  
   >> test is completed 'successfully'
 
 
@@ -21,6 +24,9 @@ then run ( It will automatically run a test on the results files ):
 **Main Usage**
 
   from BCAWT import BCAWT
+  
   BCAWT.BCAW('Ecoli','result_folder',fasta=True,Auto=True)
+  
   >> Results Saved
+  
 
