+++ docs/BCAWT.rst
@@ -5,7 +5,7 @@ BCAWT package
 BCAWT.BCAWT module
 ------------------
 
-.. automodule:: BCAWT.BCAWT
+.. automodule:: BCAWT
    :members:
    :undoc-members:
    :show-inheritance:
@@ -13,7 +13,7 @@ BCAWT.BCAWT module
 BCAWT.BCAWT\_auto\_test module
 ------------------------------
 
-.. automodule:: BCAWT.BCAWT_auto_test
+.. automodule:: BCAWT_auto_test
    :members:
    :undoc-members:
    :show-inheritance:
@@ -21,7 +21,7 @@ BCAWT.BCAWT\_auto\_test module
 BCAWT.ATCG3 module
 ------------------
 
-.. automodule:: BCAWT.ATCG3
+.. automodule:: ATCG3
    :members:
    :undoc-members:
    :show-inheritance:
@@ -30,7 +30,7 @@ BCAWT.ATCG3 module
 BCAWT.CA module
 ---------------
 
-.. automodule:: BCAWT.CA
+.. automodule:: CA
    :members:
    :undoc-members:
    :show-inheritance:
@@ -38,7 +38,7 @@ BCAWT.CA module
 BCAWT.CA\_RSCU module
 ---------------------
 
-.. automodule:: BCAWT.CA_RSCU
+.. automodule:: CA_RSCU
    :members:
    :undoc-members:
    :show-inheritance:
@@ -47,7 +47,7 @@ BCAWT.CA\_RSCU module
 BCAWT.GC123 module
 ------------------
 
-.. automodule:: BCAWT.GC123
+.. automodule:: GC123
    :members:
    :undoc-members:
    :show-inheritance:
@@ -55,7 +55,7 @@ BCAWT.GC123 module
 BCAWT.GRAVY\_AROMO module
 -------------------------
 
-.. automodule:: BCAWT.GRAVY_AROMO
+.. automodule:: GRAVY_AROMO
    :members:
    :undoc-members:
    :show-inheritance:
@@ -63,7 +63,7 @@ BCAWT.GRAVY\_AROMO module
 BCAWT.Optimal\_codon\_corr\_method module
 -----------------------------------------
 
-.. automodule:: BCAWT.Optimal_codon_corr_method
+.. automodule:: Optimal_codon_corr_method
    :members:
    :undoc-members:
    :show-inheritance:
@@ -71,7 +71,7 @@ BCAWT.Optimal\_codon\_corr\_method module
 BCAWT.P2\_index module
 ----------------------
 
-.. automodule:: BCAWT.P2_index
+.. automodule:: P2_index
    :members:
    :undoc-members:
    :show-inheritance:
@@ -79,16 +79,7 @@ BCAWT.P2\_index module
 BCAWT.PR2\_plot\_data module
 ----------------------------
 
-.. automodule:: BCAWT.PR2_plot_data
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
-
-Module contents
----------------
-
-.. automodule:: BCAWT
+.. automodule:: PR2_plot_data
    :members:
    :undoc-members:
    :show-inheritance:
