+++ docs/BCAWT.rst
@@ -1,11 +1,10 @@
 BCAWT package
 =============
 
-
 BCAWT.BCAWT module
 ------------------
 
-.. automodule:: BCAWT
+.. automodule:: BCAWT.BCAWT
    :members:
    :undoc-members:
    :show-inheritance:
@@ -13,24 +12,26 @@ BCAWT.BCAWT module
 BCAWT.BCAWT\_auto\_test module
 ------------------------------
 
-.. automodule:: BCAWT_auto_test
+.. automodule:: BCAWT.BCAWT_auto_test
    :members:
    :undoc-members:
    :show-inheritance:
    
+
 BCAWT.ATCG3 module
 ------------------
 
-.. automodule:: ATCG3
+.. automodule:: BCAWT.ATCG3
    :members:
    :undoc-members:
    :show-inheritance:
+
 
 
 BCAWT.CA module
 ---------------
 
-.. automodule:: CA
+.. automodule:: BCAWT.CA
    :members:
    :undoc-members:
    :show-inheritance:
@@ -38,7 +39,7 @@ BCAWT.CA module
 BCAWT.CA\_RSCU module
 ---------------------
 
-.. automodule:: CA_RSCU
+.. automodule:: BCAWT.CA_RSCU
    :members:
    :undoc-members:
    :show-inheritance:
@@ -47,7 +48,7 @@ BCAWT.CA\_RSCU module
 BCAWT.GC123 module
 ------------------
 
-.. automodule:: GC123
+.. automodule:: BCAWT.GC123
    :members:
    :undoc-members:
    :show-inheritance:
@@ -55,7 +56,7 @@ BCAWT.GC123 module
 BCAWT.GRAVY\_AROMO module
 -------------------------
 
-.. automodule:: GRAVY_AROMO
+.. automodule:: BCAWT.GRAVY_AROMO
    :members:
    :undoc-members:
    :show-inheritance:
@@ -63,7 +64,7 @@ BCAWT.GRAVY\_AROMO module
 BCAWT.Optimal\_codon\_corr\_method module
 -----------------------------------------
 
-.. automodule:: Optimal_codon_corr_method
+.. automodule:: BCAWT.Optimal_codon_corr_method
    :members:
    :undoc-members:
    :show-inheritance:
@@ -71,7 +72,7 @@ BCAWT.Optimal\_codon\_corr\_method module
 BCAWT.P2\_index module
 ----------------------
 
-.. automodule:: P2_index
+.. automodule:: BCAWT.P2_index
    :members:
    :undoc-members:
    :show-inheritance:
@@ -79,7 +80,7 @@ BCAWT.P2\_index module
 BCAWT.PR2\_plot\_data module
 ----------------------------
 
-.. automodule:: PR2_plot_data
+.. automodule:: BCAWT.PR2_plot_data
    :members:
    :undoc-members:
    :show-inheritance:
