+++ docs/index.rst
@@ -1,0 +1,34 @@
+Welcome to BCAW tool's documentation!
+=====================================
+
+
+Guide
+^^^^^
+.. toctree::
+   :maxdepth: 2
+   
+   Usage
+   install
+   Dependencies
+   Contribution
+   License
+   help
+   
+
+API
+^^^
+.. toctree::
+   :maxdepth: 2
+
+   
+   BCAWT
+
+
+
+
+Indices and tables
+==================
+
+* :ref:`genindex`
+* :ref:`modindex`
+* :ref:`search`
