### 2019-06-27 e878f72
Delete ATCG3.py
