### 2019-06-27 805bf6a
Delete BCAWT_auto_test.py
