### 2019-06-27 40d4ee6
Delete CA_RSCU.py
