### 2019-06-27 3cd9561
Delete BCAWT.rst
