### 2019-06-27 1d0b965
Delete Contribution.rst
