### 2019-06-27 8812f5a
Delete Dependencies.rst
