+++ docs/Dependencies.rst
@@ -1,15 +1,0 @@
-Dependencies
-============
-
-
-- Biopython
-
-- pandas
-
-- CAI
-
-- scipy
-
-- matplotlib
-
-- numpy
