### 2019-06-27 a24f528
Delete GC123.py
