### 2019-06-27 d4f1382
Delete License.rst
