### 2019-06-27 121c6c9
Delete modules.rst
