+++ docs/modules.rst
@@ -1,7 +1,0 @@
-BCAWT
-=====
-
-.. toctree::
-   :maxdepth: 4
-
-   BCAWT
