### 2019-06-27 1a74b48
Delete contents.rst
