+++ docs/Contribution.rst
@@ -1,0 +1,5 @@
+Contribution Guidelines
+=========================
+
+To report bugs or seek support please open an issue on this repository. 
+Contributions to the software are welcome; please open an issue for further discussion.
