+++ docs/Dependencies.rst
@@ -1,0 +1,15 @@
+Dependencies
+============
+
+
+- Biopython
+
+- pandas
+
+- CAI
+
+- scipy
+
+- matplotlib
+
+- numpy
