+++ Source Code/GRAVY_AROMO.py
@@ -1,4 +1,20 @@
 def GRAvy_ARomo(seq , G = False , A = False):
+    """calculating Gravy and Aroma.
+
+    Args:
+        seq (str):DNA sequence
+        G (bool): default = False
+        A (bool): default = False
+        
+
+    Returns:
+        - Gravy value if arg(G) is True
+		
+        - Aroma value if arg(A) is True
+		
+        - None if both args are False
+
+    """
     from Bio.SeqUtils.ProtParam import ProteinAnalysis
     from Bio.Seq import Seq
     translate_seq = str(seq.translate(table=1))

[…]
