+++ docs/conf.py
@@ -13,7 +13,7 @@
 import os
 import sys
 # sys.path.insert(0, os.path.abspath('.'))
-sys.path.insert(0, os.path.abspath("../CAI"))
+sys.path.insert(0, os.path.abspath("../Source Code"))
 
 
 # -- Project information -----------------------------------------------------
