+++ docs/BCAWT.rst
@@ -1,102 +1,88 @@
 BCAWT package
 =============
 
-Submodules
-----------
-
-BCAWT.ATCG3 module
-------------------
-
-.. automodule:: BCAWT.ATCG3
-   :members:
-   :undoc-members:
-   :show-inheritance:
-
 BCAWT.BCAWT module
 ------------------
 
 .. automodule:: BCAWT.BCAWT
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 BCAWT.BCAWT\_auto\_test module
 ------------------------------
 
 .. automodule:: BCAWT.BCAWT_auto_test
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
+
+BCAWT.ATCG3 module
+------------------
+
+.. automodule:: BCAWT.ATCG3
+   :members:
+    
+    
+
+
 
 BCAWT.CA module
 ---------------
 
 .. automodule:: BCAWT.CA
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 BCAWT.CA\_RSCU module
 ---------------------
 
 .. automodule:: BCAWT.CA_RSCU
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
-BCAWT.ENc module
-----------------
 
-.. automodule:: BCAWT.ENc
-   :members:
-   :undoc-members:
-   :show-inheritance:
 
 BCAWT.GC123 module
 ------------------
 
 .. automodule:: BCAWT.GC123
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 BCAWT.GRAVY\_AROMO module
 -------------------------
 
 .. automodule:: BCAWT.GRAVY_AROMO
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 BCAWT.Optimal\_codon\_corr\_method module
 -----------------------------------------
 
 .. automodule:: BCAWT.Optimal_codon_corr_method
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 BCAWT.P2\_index module
 ----------------------
 
 .. automodule:: BCAWT.P2_index
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 BCAWT.PR2\_plot\_data module
 ----------------------------
 
 .. automodule:: BCAWT.PR2_plot_data
    :members:
-   :undoc-members:
-   :show-inheritance:
+    
+    
 
 
-Module contents
----------------
-
-.. automodule:: BCAWT
-   :members:
-   :undoc-members:
-   :show-inheritance:
