+++ docs/BCAWT.rst
@@ -12,7 +12,7 @@ BCAWT.BCAWT module
 BCAWT.BCAWT\_auto\_test module
 ------------------------------
 
-.. automodule:: BCAWT.BCAWT_auto_test
+.. automodule:: BCAWT_auto_test
    :members:
     
     
@@ -20,17 +20,15 @@ BCAWT.BCAWT\_auto\_test module
 BCAWT.ATCG3 module
 ------------------
 
-.. automodule:: BCAWT.ATCG3
+.. automodule:: ATCG3
    :members:
     
-    
-
 
 
 BCAWT.CA module
 ---------------
 
-.. automodule:: BCAWT.CA
+.. automodule:: CA
    :members:
     
     
@@ -38,17 +36,15 @@ BCAWT.CA module
 BCAWT.CA\_RSCU module
 ---------------------
 
-.. automodule:: BCAWT.CA_RSCU
+.. automodule:: CA_RSCU
    :members:
     
-    
-
 
 
 BCAWT.GC123 module
 ------------------
 
-.. automodule:: BCAWT.GC123
+.. automodule:: GC123
    :members:
     
     
@@ -56,7 +52,7 @@ BCAWT.GC123 module
 BCAWT.GRAVY\_AROMO module
 -------------------------
 
-.. automodule:: BCAWT.GRAVY_AROMO
+.. automodule:: GRAVY_AROMO
    :members:
     
     
@@ -64,15 +60,14 @@ BCAWT.GRAVY\_AROMO module
 BCAWT.Optimal\_codon\_corr\_method module
 -----------------------------------------
 
-.. automodule:: BCAWT.Optimal_codon_corr_method
+.. automodule:: Optimal_codon_corr_method
    :members:
-    
     
 
 BCAWT.P2\_index module
 ----------------------
 
-.. automodule:: BCAWT.P2_index
+.. automodule:: P2_index
    :members:
     
     
@@ -80,7 +75,7 @@ BCAWT.P2\_index module
 BCAWT.PR2\_plot\_data module
 ----------------------------
 
-.. automodule:: BCAWT.PR2_plot_data
+.. automodule:: PR2_plot_data
    :members:
     
     
