+++ docs/Usage.rst
@@ -10,12 +10,12 @@ https://github.com/AliYoussef96/BCAWTool/blob/master/Ecoli.fasta
 then run ( It will automatically run a test on the results files ):
 
 	from BCAWT import BCAWT_auto_test
-    
-    BCAWT_auto_test.auto_test()
+	
+	BCAWT_auto_test.auto_test()
     
 	BCAWT_auto_test.auto_check_files()
-    
-    >> test is completed 'successfully'
+	
+	>> test is completed 'successfully'
     
 **Main Usage**
 
@@ -23,4 +23,4 @@ then run ( It will automatically run a test on the results files ):
     
 	BCAWT.BCAW('Ecoli','result_folder',fasta=True,Auto=True)
 	
-    >> Results Saved
+    	>> Results Saved
