+++ requirements.txt
@@ -1,0 +1,6 @@
+Biopython
+pandas
+CAI
+scipy
+matplotlib
+numpy
