### 2019-06-27 abb28fc
Update install.rst
