+++ docs/install.rst
@@ -3,6 +3,6 @@ Installation
 
 *Using pip*
 
-	pip install
+	pip install BCAWT
 
 or by downloading the source code from https://github.com/AliYoussef96/BCAW-Tool
