+++ docs/Statement.rst
@@ -1,0 +1,9 @@
+Statement of Need
+==================
+There are no tools available enable users to run a whole automated workflow for codon usage bias analysis. 
+
+Using python 3.7 BCAW Tool ( Bio Codon Analysis Workflow Tool ) was developed to address this problem. 
+
+Expected results from BCAW Tool are; nucleotide content, various codon usage measures (as Effective number of codons, Codon adaptation index, etc.) as well as correlation analysis with graphs. Further, Correspondence Analysis and determination of putative optimal codons are included. BCAW Tool manages a complete automated workflow to analysis the codon usage bias for genes and genomes of any organism. BCAW Tool is available as executable application, work under Windows operating system, also a source code is available.
+
+For more details about CUB, and the equations used in BCAWT see.
