+++ docs/Statement.rst
@@ -6,4 +6,4 @@ Using python 3.7 BCAW Tool ( Bio Codon Analysis Workflow Tool ) was developed to
 
 Expected results from BCAW Tool are; nucleotide content, various codon usage measures (as Effective number of codons, Codon adaptation index, etc.) as well as correlation analysis with graphs. Further, Correspondence Analysis and determination of putative optimal codons are included. BCAW Tool manages a complete automated workflow to analysis the codon usage bias for genes and genomes of any organism. BCAW Tool is available as executable application, work under Windows operating system, also a source code is available.
 
-For more details about CUB, and the equations used in BCAWT see.
+For more details about CUB, and the equations used in BCAWT https://github.com/AliYoussef96/BCAW-Tool/blob/master/Introduction%20to%20codon%20usage%20bias.pdf.
