### 2019-06-27 bf2d5ae
Update paper.md
