### 2019-06-27 32c69ea
Update paper.md
