### 2019-06-27 9e66665
Create CONTRIBUTING.md
