### 2019-06-27 85c7d43
Update CONTRIBUTING.md
