+++ CONTRIBUTING.md
@@ -1,6 +1,6 @@
 # Contributions to the software are welcome.
 
-For bugs and suggestions, the most effective way to reach the team is by raising an issue on the github issue tracker. 
+For bugs and suggestions, the most effective way is by raising an issue on the github issue tracker. 
 Github allows you to classify your issues so that we know if it is a bug report, feature request or feedback to the authors.
 
 If you wish to contribute some changes to the code then you should submit a [pull request](https://github.com/AliYoussef96/BCAW-Tool/pulls)
