+++ Source Code/BCAWT.py
@@ -5,19 +5,25 @@
     #2- fasta file for refrence set
 
 ########main file
-def BCAW(input_the_main_fasta_file,save_folder_name,input_the_ref_fasta_file=None,fasta = False,txt=False,Auto=False):
+def BCAW(input_the_main_fasta_file,save_folder_name,input_the_ref_fasta_file=None,genetic_code_ = 1, fasta = False,txt=False,Auto=False):
     """
     BCAWT ( Bio Codon Analysis Workflow Tool ), it manages a complete workflow to analysis
     the codon usage bias for genes and genomes of any organism..
 
     Args:
 
-        input_the_main_fasta_file (str): fasta file contains DNA sequence ( don't enter it with .fasta )
-        or text file contains paths for fasta files ( don't enter it with .txt )
+        input_the_main_fasta_file (str): fasta file contains DNA sequence ( don't enter it with .fasta ) or text file contains paths for fasta files ( don't enter it with .txt )
+
         save_folder_name (str): folder name where the result will be saved
+
         input_the_ref_fasta_file (str): fasta file contains reference DNA sequence, default = None
+
+        genetic_code_(int): default = 1, The Genetic Codes number described by NCBI (https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi)
+
         fasta (bool): default = False, if the first arg (input_the_main_fasta_file) is text file.
+
         txt (bool): default = False, if the first arg (input_the_main_fasta_file) is fasta file.
+
         Auto (bool): default = False, if input_the_ref_fasta_file not None.
 
     Notes:
@@ -28,10 +34,23 @@ def BCAW(input_the_main_fasta_file,save_folder_name,input_the_ref_fasta_file=Non
         - Auto (bool): should be = True to auto-generate a reference set, when arg (input_the_ref_fasta_file) not available ( = None )
 
     Returns:
-       github.com/AliYoussef96/BCAW-Tool/blob/master/Table.png
+       for details see: github.com/AliYoussef96/BCAW-Tool/blob/master/Table.png
 
 
     """
+
+    #bad arguments raise errors
+
+    if fasta == True and txt == True:
+        raise TypeError("Two file extensions are specified, ( fasta = True, txt = True ). Only one is allowed.")
+    if input_the_ref_fasta_file != None and Auto == True:
+        raise TypeError("Both genes reference set and Auto are specified  Only one is allowed.")
+    if txt == False and fasta == False :
+        raise TypeError ("The file extension is not specified, (fasta = False, txt = False )")
+
+
+
+
 
     import Bio
     from Bio import SeqIO
@@ -82,9 +101,8 @@ def BCAW(input_the_main_fasta_file,save_folder_name,input_the_ref_fasta_file=Non
             i = i.replace('\n','')
             i = i + ".fasta"
             file_name_list.append(i)
-    else:
-        print  ('"Something going wrong!!!, Please review the tutorial and your input and try again"')
-        return
+
+
     ############ref file
     file_name_ref_list = []
     if input_the_ref_fasta_file != None and fasta == True and Auto == False:
@@ -129,9 +147,7 @@ def BCAW(input_the_main_fasta_file,save_folder_name,input_the_ref_fasta_file=Non
                 all_ref_seq.append(seq_main_seq_modifi_)
     elif Auto and Auto == True and input_the_ref_fasta_file == None:
         list_ref = []
-    else:
[…]
