+++ docs/Usage.rst
@@ -25,5 +25,6 @@ then run ( It will automatically run a test on the results files ):
 	
     	>> Results Saved
 
- 
+**Important Note:** BCAW tool expect coding sequences 
+
 For more information on the equations used in the tool https://github.com/AliYoussef96/BCAW-Tool/blob/master/Introduction%20to%20codon%20usage%20bias.pdf
