+++ docs/Documentations.rst
@@ -1,11 +1,8 @@
 Documentations
 ================
 
+1- For more information about codon usage bias (CUB) and equations used to analyze CUB in the BCAW tool >> https://github.com/AliYoussef96/BCAW-Tool/blob/master/Introduction%20to%20codon%20usage%20bias.pdf
 
-- For more information about codon usage bias (CUB) and equations used to analyze CUB in the BCAW tool >> https://github.com/AliYoussef96/BCAW-Tool/blob/master/Introduction%20to%20codon%20usage%20bias.pdf
+2- For more information about the BCAW tool and the API >> https://bcaw-tools-documentation.readthedocs.io/en/latest/index.html
 
-
-- For more information about the BCAW tool and the API >> https://bcaw-tools-documentation.readthedocs.io/en/latest/index.html
-
-
-- For more information about the description output.
+3- For more information about the description output.
