+++ Outputs description.md
@@ -1,0 +1,18 @@
+## Abbreviations list
+
+
+|Abbreviation|Description|
+|------------|-----------|
+| GC | G and C nucleotides content |
+| GC1 | G and C nucleotides content at first codon postion |
+| GC2 | G and C nucleotides content at second codon postion |
+| GC3 | G and C nucleotides content at third codon postion |
+| GC12 | G and C nucleotides content at first + second codon postions |
+| AT | A and T nucleotides content |
+| AT3  | A and T nucleotides content  at third codon position |
+| A3 | A nucleotides content  at third codon position |
+| T3 | T3 nucleotides content  at third codon position |
+| C3 | C3 nucleotides content  at third codon position |
+| G3 | G3 nucleotides content  at third codon position |
+| GRAVY | General average hydropathicity |
+| AROMO | Aromaticity |
