### 2019-06-28 1a14b29
Update Outputs description.md
