+++ Plots/Plots Description.md
@@ -1,1 +1,43 @@
-#
+![fig1: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_AROMA%20Vs.%20GC.png)
+
+Fig1: Relationship between GC overall content and the aromaticity. Blue dots represent the genes
+
+![fig2: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_CAI%20Vs.%20Gene%20Length.png)
+
+Fig2: Relationship between the CAI and the gene length. Blue dots represent the genes
+
+![fig3: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_ACTG3%20violin%20plot.png)
+
+Fig3: ATCG3 violin plot: Explain the range of A3, T3, C3 and, G3 content as a normal box plot but also show the probability density of the data at different values.
+
+![fig4: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_CA_RSCU.csv_CA_gens_plot.png)
+
+Fig4: Correspondence analysis plot each genes. Axis 1 and Axis 2 represent the largest contributors to the CUB values of genes in the correspondence analysis and the blue dots represent the genes.
+
+![fig5: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_CA_RSCU_CA_codos_plot.png)
+
+Fig5: Correspondence analysis plot for Escherichia coli each codons. Axis 1 and Axis 2 represent the largest contributors to the RSCU values of genes in the correspondence analysis and the blue dots represent the codons.
+
+![fig6: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_ENc%20Vs.%20CAI.png)
+
+Fig6: Relationship between the CAI and the ENc. Blue dots represent the genes.
+
+![fig7: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_ENc%20Vs.%20GC3.png)
+
+Fig7: Relationship between the ENc and the GC3 content at third codon position. Blue dots represent the genes.
+
+![fig8: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_GC%20violin%20plot.png)
+
+Fig8: Explain the range of GC, GC1, GC2 and, GC3 content as a normal box plot but also show the probability density of the data at different values.
+
+![fig9: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_GC3%20Vs.%20GC12.png)
+
+Fig9: Neutrality Plot, Relationship between the GC3 and the GC12 content at third codon position. Blue dots represent the genes.
+
+![fig10: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_GRAVY%20Vs.%20GC.png)
+
+Fig10: Relationship between GC overall content and (Gravy). Blue dots represent the genes.
+
+![fig11: ](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_PR2-plot.png)
+
+Fig11: Relationship between AT-bias and GC-bias. Blue dots represent the genes.
