### 2019-06-28 #1500 @AliYoussef96
Hi @mgymrek

**Install process**

For [setup.py](https://github.com/AliYoussef96/BCAW-Tool/blob/master/setup.py) is available now, as well as installation using pip.

```
pip install BCAWT
```
**Usage/docs**

1- BCAWT can now take all the arguments at once.

```
from BCAWT import BCAWT
BCAWT.BCAW('Ecoli','result_folder',genetic_code_=11,fasta=True,Auto=True)
```
2- Inputs, inputs, and how to obtain such fasta files are now on README.
As well as I added [Abbreviations table](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Abbreviations.md), Also, I added an introduction to codon usage bias and equations used in the BCAW tool with some useful references [here](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Introduction%20to%20codon%20usage%20bias.pdf).

3- All plots are described [here](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Plots%20Description.md).

**Testing**

For the BCAW tool, I did BCAWT_auto_test.py. I did describe how to use this test in README.md and in the documentation.

Also, I wrote [test.py](https://github.com/AliYoussef96/BCAW-Tool/blob/master/test.py) to tests functions with automated tests

**paper**

I am working on a summary to describe in more detail what the outputs are.

**paper**
