### 2019-06-28 7f9718f
Create plots.rst
