### 2019-06-29 7766895
Update GC123.py
