+++ README.md
@@ -70,11 +70,11 @@ BCAWT.BCAW('Ecoli','result_folder',genetic_code_=11,fasta=True,Auto=True)
 ```
 ### Input
 
-input_the_main_fasta_file (str) : fasta file contains DNA sequence 
+**input_the_main_fasta_file (str)** : fasta file contains DNA sequence 
 
-input_the_ref_fasta_file (str) : fasta file contains reference DNA sequence, default = None. Here Auto is True, to generate automatically reference genes set
+**input_the_ref_fasta_file (str)** : fasta file contains reference DNA sequence, default = None. Here Auto is True, to generate automatically reference genes set
 
-genetic_code_ (int) : default = 1, The Genetic Codes number described by [NCBI](https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi)
+**genetic_code_ (int)** : default = 1, The Genetic Codes number described by [NCBI](https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi)
 
 **Important Note:** BCAW tool expect coding sequences
 
