### 2019-06-29 bf10f38
Update paper.md
