+++ paper/paper.md
@@ -33,10 +33,8 @@ BCAWT.BCAW('Ecoli','result_folder',genetic_code_=11,fasta=True,Auto=True)
 
 The expected outputs are 23 different files, 8 are in comma-separated values ( CSV ) format, 4 are in text format and, 11 graphs in portable network graphics ( PNG ) format [(Table 1)](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Table.png). The equations used for analyzing CUB in the tool and the expected outputs, as well as the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
 
-The advantages for BCAWtool over the existing tools are; 1- Automated Workflow,  2- No limitations on the number of genes, 3- The method used to determine optimal codons named correlation method is only available in the BCAW tool,  4- Plots as violin plots for nucleotides contents (Fig. 1), and other plots are generated by the BCAWtool and no need for using another software to draw such plots.
+The advantages for BCAWtool over the existing tools are; 1- Automated Workflow,  2- No limitations on the number of genes, 3- The method used to determine optimal codons named correlation method is only available in the BCAW tool,  4- Plots as violin plots for nucleotides contents, and other plots are generated by the BCAWtool and no need for using another software to draw such plots.
 
-![(Fig 1)](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Ecoli.fasta_ACTG3%20violin%20plot.png)
 
-Fig. 1: ATCG3 violin plot: Explain the range of A3, T3, C3 and, G3 content at third codon positions as a normal box plot but also show the probability density of the data at different values.
 
 # References
