+++ paper/paper.md
@@ -31,9 +31,31 @@ from BCAWT import BCAWT
 BCAWT.BCAW('Ecoli','result_folder',genetic_code_=11,fasta=True,Auto=True)
 ```
 
-The expected outputs are 23 different files, 8 are in comma-separated values ( CSV ) format, 4 are in text format and, 11 graphs in portable network graphics ( PNG ) format [(Table 1)](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Table.png). The equations used for analyzing CUB in the tool and the expected outputs, as well as the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
+The expected outputs are 23 different files, 8 are in comma-separated values ( CSV ) format, 4 are in text format and, 11 graphs in portable network graphics ( PNG ) format. The equations used for analyzing CUB in the tool and the expected outputs, as well as the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
 
 The advantages for BCAWtool over the existing tools are; 1- Automated Workflow,  2- No limitations on the number of genes, 3- The method used to determine optimal codons named correlation method is only available in the BCAW tool,  4- Plots as violin plots for nucleotides contents, and other plots are generated by the BCAWtool and no need for using another software to draw such plots.
+
+# OutPut summary
+
+Visualization is the best way to present data. Hence, BCAWtool returns CSV files contain the codon usage analysis and the codon usage bias indices output ( Table 1 ), after that, it returns 11 plot  (Fig 1), so one could interpret the results easily.
+
+|CSV file name|Description|
+|------------|-----------|
+| ATCG | contains ; gene id, GC, GC1, GC2, GC3, GC12, AT, AT3    A3, T3, C3, G3, GRAVY, AROMO and, Gene Length |
+| CA_RSCU | contains ; each RSCU result for each codon in each genes |
+| CA_RSCUcodons | contains ; correspondence analysis first 4 axis for each codon |
+| CA_RSCUgenes | contains ; correspondence analysis first 4 axis for each gene |
+| CAI | contains ; gene id and CAI index |
+| ENc | contains ; gene id and ENc index. |
+| P2-index | contains ; gene id and P2 index |
+| optimal codons | contains; putative optimal codons detected |
+
+Table 1: Explain the CSV files output from BCAWtool
+
+
+![Fig 1](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/All%20plots.jpg)
+
+Fig 1: All output plots from BCAWtool analysis for coding sequence from Escherichia coli.
 
 
 
