+++ paper/paper.md
@@ -39,6 +39,8 @@ The advantages for BCAWtool over the existing tools are; 1- Automated Workflow, 
 
 Visualization is the best way to present data. Hence, BCAWtool returns CSV files contain the codon usage analysis and the codon usage bias indices output ( Table 1 ), after that, it returns 11 plot  (Fig 1), so one could interpret the results easily.
 
+**Table 1: Explain the CSV files output from BCAWtool**
+
 |CSV file name|Description|
 |------------|-----------|
 | ATCG | contains ; gene id, GC, GC1, GC2, GC3, GC12, AT, AT3    A3, T3, C3, G3, GRAVY, AROMO and, Gene Length |
@@ -50,12 +52,10 @@ Visualization is the best way to present data. Hence, BCAWtool returns CSV files
 | P2-index | contains ; gene id and P2 index |
 | optimal codons | contains; putative optimal codons detected |
 
-Table 1: Explain the CSV files output from BCAWtool
-
+**Fig 1: All output plots from BCAWtool analysis for coding sequence from Escherichia coli**
 
 ![Fig 1](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/All%20plots.jpg)
 
-Fig 1: All output plots from BCAWtool analysis for coding sequence from Escherichia coli.
 
 
 
