+++ paper/paper.md
@@ -35,7 +35,7 @@ The expected outputs are 23 different files, 8 are in comma-separated values ( C
 
 The advantages for BCAWtool over the existing tools are; 1- Automated Workflow,  2- No limitations on the number of genes, 3- The method used to determine optimal codons named correlation method is only available in the BCAW tool,  4- Plots as violin plots for nucleotides contents, and other plots are generated by the BCAWtool and no need for using another software to draw such plots.
 
-# OutPut summary
+# Output summary
 
 Visualization is the best way to present data. Hence, BCAWtool returns CSV files contain the codon usage analysis and the codon usage bias indices output ( Table 1 ), after that, it returns 11 plot  (Fig 1), so one could interpret the results easily.
 
