### 2019-07-04 #1500 @AliYoussef96
Hi @mgymrek

For the [ paper](https://github.com/AliYoussef96/BCAW-Tool/blob/master/paper/paper.md) Output summary, I add a table summary, as well as a figure, shows all the output plots.

I'm Waiting for your feedback.

Hi @Benjamin-Lee

I finished working on the issues you added. And I'm Waiting for your feedback.
