### 2019-07-06 955a1d1
Update setup.py
