### 2019-07-16 #1500 @mgymrek
thanks @AliYoussef96 for the updates. The changes look good.

One comment on install: I wonder whether it is critical to require Python >=3.7? Many systems will still have earlier versions making it difficult to install and test with pip.

Also, comments on the paper:

Summary
* "Futher, Using" -> "Further, using the"
* "The tool also includes statistical analysis as" -> "The tool also includes statistical analysis such as"

Implementation
* "build in" -> "built in"
* "bunch of analysis" -> "variety of analyses"
* It is not as useful to say how many different files are output. It would more helpful to just summarize the types of output. e.g. plots showing XYZ
* "BCAWtool" -> "BCAW Tool". same elsewhere, make sure you are consistent with how you reference the tool

Output summary:
* "CSV files contain" -> "CSV file containing"
* Start a new sentence after Table 1, e.g. "It also returns 11 plots depicting output for easy interpretation (Fig. 1)"
