### 2019-07-16 e7c1498
Update paper.md
