+++ paper/paper.md
@@ -20,26 +20,26 @@ The redundancy in the genetic code means that apart from methionine and tryptoph
 
 Many measurements have been developed to analysis and study CUB; effective number of codons ( ENc ) [@Wright1990], codon adaptation index ( CAI ), relative synonymous codon usage ( RSCU ) [@Sharp1987] and, translational selection index ( P2-index ) [@Wang2018]. Also, statistical analysis has been used to investigate the effect of different factors as selection and mutation on shaping CUB as; Correspondence analysis, Parity Rule 2-plot Analysis and, Neutrality Plot [@Song2017]. BCAW tool was developed to analyze such phenomena ( Codon Usage Bias ) by the aforementioned  measurements.
 
-Various tools are available to analyze and measure CUB, but they lack some important measurements and plots for CUB analysis. What BCAWtool does is an automated workflow to study the CUB of any organism genes by all the measurements and plots mentioned above. Further, Using correlation method to determine the optimal codons described by [@Hershberg2009] is implemented for the first time in BCAW tool. The tool also includes statistical analysis as correspondence analysis, correlation analysis, and t-test.
+Various tools are available to analyze and measure CUB, but they lack some important measurements and plots for CUB analysis. What BCAW tool does is an automated workflow to study the CUB of any organism genes by all the measurements and plots mentioned above. Further, using the correlation method to determine the optimal codons described by [@Hershberg2009] is implemented for the first time in BCAW tool. The tool also includes statistical analysis such as correspondence analysis, correlation analysis, and t-test.
  
 # Implementation
 
-BCAWtool was developed using python 3.7 with build in and third-party packages [@Lee2018]. The usage of BCAW Tool is made to be very easy where users have only to input a fasta format file containing genes to be analyzed, and a bunch of analysis will be performed.
+BCAW tool was developed using python 3.7 with built in and third-party packages [@Lee2018]. The usage of BCAW Tool is made to be very easy where users have only to input a fasta format file containing genes to be analyzed, and variety of analyses will be performed.
 
 ```
 from BCAWT import BCAWT
 BCAWT.BCAW('Ecoli','result_folder',genetic_code_=11,fasta=True,Auto=True)
 ```
 
-The expected outputs are 23 different files, 8 are in comma-separated values ( CSV ) format, 4 are in text format and, 11 graphs in portable network graphics ( PNG ) format. The equations used for analyzing CUB in the tool and the expected outputs, as well as the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
+The expected outputs from the BCAW tool could be divided into three groups. The first one is data in CSV format described in (Table 1), the second group is plots summarized in (Fig 1), and the last one is text files each contains a different result for a different statistical test. The equations used for analyzing CUB in the tool, and the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
 
-The advantages for BCAWtool over the existing tools are; 1- Automated Workflow,  2- No limitations on the number of genes, 3- The method used to determine optimal codons named correlation method is only available in the BCAW tool,  4- Plots as violin plots for nucleotides contents, and other plots are generated by the BCAWtool and no need for using another software to draw such plots.
[…]
