### 2019-07-16 #1500 @AliYoussef96
Thanks, @mgymrek for your comments.

**1. Comment on install to require Python >=3.7**.
[(README)](https://github.com/AliYoussef96/BCAW-Tool/blob/master/README.md).

**2. Updates to the paper**

**Summary**

- The summary section is updated.

**Implementation**

- The Implementation section is updated. For "BCAWtool" I converted it to "BCAW tool" everywhere in the paper and also in the readme.

**Output summary**

- The Output summary section is updated.

Waiting for further comments.
