### 2019-07-17 #1500 @AliYoussef96
Hi @Benjamin-Lee

Can we close the opened issues now?
