+++ .travis.yml
@@ -1,0 +1,8 @@
+dist: xenial
+language: python
+python:
+  - "3.7"
+install:
+  - pip install -r requirements.txt
+script:
+  - test
