+++ .travis.yml
@@ -4,5 +4,6 @@ python:
   - "3.7"
 install:
   - pip install -r requirements.txt
+os: windows
 script:
   - pytest
