+++ .travis.yml
@@ -1,8 +1,8 @@
-dist: xenial
+dist: bionic
 language: python
 python:
   - "3.7"
 install:
   - pip install -r requirements.txt
 script:
-  - test.py
+  - pytest
