+++ .travis.yml
@@ -5,4 +5,4 @@ python:
 install:
   - pip install -r requirements.txt
 script:
-  - pytest
+  - pytest test.py
