### 2019-07-23 65c9463
Delete ATCG3.py
