### 2019-07-23 1747b27
Delete BCAWT.py
