+++ Source Code/BCAWT.py
@@ -1,602 +1,0 @@
-
-#### the expected input are CDS
-#input needed
-    #1- fasta file for CDS need to be analyze ( -s for single fasta -m for multifily))
-    #2- fasta file for refrence set
-
-########main file
-def BCAW(input_the_main_fasta_file,save_folder_name,input_the_ref_fasta_file=None,genetic_code_ = 1, fasta = False,txt=False,Auto=False):
-    """
-    BCAWT ( Bio Codon Analysis Workflow Tool ), it manages a complete workflow to analysis
-    the codon usage bias for genes and genomes of any organism..
-
-    Args:
-
-        input_the_main_fasta_file (str): fasta file contains DNA sequence ( don't enter it with .fasta ) or text file contains paths for fasta files ( don't enter it with .txt )
-
-        save_folder_name (str): folder name where the result will be saved
-
-        input_the_ref_fasta_file (str): fasta file contains reference DNA sequence, default = None
-
-        genetic_code_(int): default = 1, The Genetic Codes number described by NCBI (https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi)
-
-        fasta (bool): default = False, if the first arg (input_the_main_fasta_file) is text file.
-
-        txt (bool): default = False, if the first arg (input_the_main_fasta_file) is fasta file.
-
-        Auto (bool): default = False, if input_the_ref_fasta_file not None.
-
-    Notes:
-        - fasta (bool): should be = True when the first arg (input_the_main_fasta_file) is fasta file.
-
-        - txt (bool): should be = True when the first arg (input_the_main_fasta_file) is text file.
-
-        - Auto (bool): should be = True to auto-generate a reference set, when arg (input_the_ref_fasta_file) not available ( = None )
-
-    Returns:
-       for details see: github.com/AliYoussef96/BCAW-Tool/blob/master/Table.png
-
-
-    """
-
-    #bad arguments raise errors
-
-    if fasta == True and txt == True:
-        raise TypeError("Two file extensions are specified, ( fasta = True, txt = True ). Only one is allowed.")
-    if input_the_ref_fasta_file != None and Auto == True:
-        raise TypeError("Both genes reference set and Auto are specified  Only one is allowed.")
-    if txt == False and fasta == False :
-        raise TypeError ("The file extension is not specified, (fasta = False, txt = False )")
-
-
-
-
-
-    import Bio
-    from Bio import SeqIO
-    from Bio.Alphabet import IUPAC
-    from Bio.Seq import Seq
-    from Bio.SeqUtils import GC
-    from Bio.Data import CodonTable
-    from Bio.Alphabet import generic_dna
-    import pandas as pd
-    from pandas import DataFrame
-    from BCAWT import GC123
-    from BCAWT import ATCG3
-    from BCAWT import ENc
-    from CAI import CAI
-    import scipy
-    from scipy import stats
-    from BCAWT import PR2_plot_data
-    from BCAWT import P2_index
-    import os
-    import matplotlib.pyplot as plt
-    import numpy as np
-    from BCAWT import CA_RSCU
-    from BCAWT import CA
-    from Bio.Data import CodonTable
-    from BCAWT import Optimal_codon_corr_method
-    import scipy.optimize as opt
-    import warnings
-    import time
-    from BCAWT import GRAVY_AROMO
[…]
