### 2019-07-23 c6fcb98
Delete BCAWT_auto_test.py
