### 2019-07-23 b72d556
Delete GC123.py
