### 2019-07-23 f5e44fa
Delete GRAVY_AROMO.py
