### 2019-07-23 76c58e6
Delete Optimal_codon_corr_method.py
