### 2019-07-23 ff45596
Delete PR2_plot_data.py
