+++ docs/install.rst
@@ -4,5 +4,3 @@ Installation
 *Using pip*
 
 	pip install BCAWT
-
-or by downloading the source code from https://github.com/AliYoussef96/BCAW-Tool
