+++ docs/intro.rst
@@ -1,0 +1,323 @@
+Introduction to Codon Usage Bias
+================================
+
+   During the translation process from mRNAs to proteins, information is
+   transmitted in the form of triple nucleotides named codons encode.
+   Amino acids are degenerate, have more than one codon to be encoded
+   except for methionine (Met) and tryptophan (Trp), thus codons encode
+   one amino acid are known as synonymous codons. Many studies on
+   different organisms reported that synonymous codons are not used
+   uniformly within and between genes of one genome, this phenomenon
+   called \`synonymous codon usage bias (SCUB) or Codon usage bias (CUB)
+   [1–4]. Further, the degree of the unequal use of synonymous codons
+   differ between species [5,6]. Hence each organism has its optimal
+   codons, an optimal codon is defined as the codon which is more
+   frequently used in highly expressed genes than in the low expressed
+   genes [7]. Two main factors shaping the codon usage of an organism
+   are mutation and selection [8–10] , other factors also described to
+   influence CUB of an organism as nucleotide composition [11],
+   synonymous substitution rate [12], tRNA abundance [13], codon
+   hydropathy and DNA replication initiation site [14], gene length [15]
+   and expression level [16]. Investigating the CUB of an organism could
+   tell about genes molecular evolution, expression, and host-pathogen
+   coadaptation. Further in biotechnology applications CUB and predicted
+   optimal codons could help to design highly expressed genes and
+   constructing suitable cloning vectors [17,18].
+
+Equations used for codon usage bias analysis
+============================================
+
+Effective number of codons
+~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+   The The effective number of codons (ENc) measures the bias of using a
+   smaller subset of codons apart from the equal use of synonymous
+   codons. In addition, the ENc measures the codon usage unbalance among
+   genes, where an amino acid is encoded by one codon in a gene would be
+   biased and negatively correlated with the ENc value. this measureENc
+   ranges from 20-61 with higher values indicating more codons being
+   used for each amino acid i.e less bias and vice versa, ENcthis
+   measures codon bias irrespective to gene length, it can be an
+   indicator of codon usage with reference to mutational bias [19]. ENc
+   was calculated using the equation given by [20]:
+
+   𝑚
+
+   𝑛 + 1
+
+𝐹𝐶𝐹
+
+   = ∑ ( 𝑖 )
+
+𝑛 + 𝑚
+-----
+
+   𝑖=1
+
+   Then ENc could be calculated by:
+
+1
+-
+
+   𝑁𝑐.𝐶𝐹 =
+
+   𝐶𝐹
+
+   Where n\ :sub:`i` is the count of codon i in m amino acid family and
+   m is the number of codons in an amino acid family
+
+Codon adaptation index
+~~~~~~~~~~~~~~~~~~~~~~
+
+   Codon adaptation index (CAI) uses a reference set of highly expressed
+   genes (e.g. ribosomal genes), this measure is an indicator of gene
+   expression levels and natural selection; it ranges from 0 to 1 with
+   higher values indicating stronger bias with respect to the reference
[…]
