### 2019-07-23 e6221c3
Update paper.md
