### 2019-07-23 98f51d0
Update intro.rst
