### 2019-07-23 c8f99d1
Update Dependencies.rst
