+++ docs/Dependencies.rst
@@ -1,5 +1,5 @@
-Dependencies
-============
+**Dependencies**
+================
 
 
 - Biopython
