### 2019-07-23 38a34c8
Update License.rst
