### 2019-07-23 982945d
Update Statement.rst
