### 2019-07-23 e13e303
Update help.rst
