+++ docs/help.rst
@@ -1,3 +1,3 @@
-Help
-========
+**Help**
+==========
 If you need more help please contact ali.mo.anwar@std.agr.cu.edu.eg
