+++ docs/install.rst
@@ -1,5 +1,5 @@
-Installation
-===============
+**Installation**
+==================
 
 *Using pip*
 
