### 2019-07-23 d398cdc
Update plots.rst
