### 2019-07-23 16af221
Update Dependencies.rst
