+++ docs/Dependencies.rst
@@ -1,8 +1,6 @@
 **Dependencies**
 ================
 
-- Python >=3.7
-
 - biopython
 
 - pandas
