+++ docs/Table_output.rst
@@ -1,0 +1,46 @@
+**Outputs**
+=============
+
+
+**The expected CSV files output**
+----------------------------------
+
+
++------------------------------------+---------------------------------+
+| CSV file name                      | Description                     |
++====================================+=================================+
+| ATCG                               | contains ; gene id, GC, GC1,    |
+|                                    | GC2, GC3, GC12, AT, AT3 A3, T3, |
+|                                    | C3, G3, GRAVY, AROMO and, Gene  |
+|                                    | Length                          |
++------------------------------------+---------------------------------+
+| CA_RSCU                            | contains ; each RSCU result for |
+|                                    | each codon in each genes        |
++------------------------------------+---------------------------------+
+| CA_RSCUcodons                      | contains ; correspondence       |
+|                                    | analysis first 4 axis for each  |
+|                                    | codon                           |
++------------------------------------+---------------------------------+
+| CA_RSCUgenes                       | contains ; correspondence       |
+|                                    | analysis first 4 axis for each  |
+|                                    | gene                            |
++------------------------------------+---------------------------------+
+| CAI                                | contains ; gene id and CAI      |
+|                                    | index                           |
++------------------------------------+---------------------------------+
+| ENc                                | contains ; gene id and ENc      |
+|                                    | index.                          |
++------------------------------------+---------------------------------+
+| P2-index                           | contains ; gene id and P2 index |
++------------------------------------+---------------------------------+
+| optimal codons                     | contains; putative optimal      |
+|                                    | codons detected                 |
++------------------------------------+---------------------------------+
+
+
+**All output plots**
+========================
+
+.. image:: https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/All%20plots.jpg
+   :width: 600
+   :alt: All output plots from BCAW tool analysis for coding sequence from Escherichia coli 
