### 2019-07-23 0dd9308
Update plots.rst
