+++ docs/Table_output.rst
@@ -1,9 +1,9 @@
 **Outputs**
-=============
+===============
 
 
 **The expected CSV files output**
-----------------------------------
+------------------------------------
 
 
 +------------------------------------+---------------------------------+
@@ -39,7 +39,7 @@
 
 
 **All output plots**
-========================
+-------------------------
 
 .. image:: https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/All%20plots.jpg
    :width: 600
