+++ docs/Documentations.rst
@@ -3,6 +3,8 @@
 
 1. For more information about the codon usage bias (CUB), equations used to analyze CUB in the BCAW tool and the API >> [BCAW tool’s documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/index.html).
 
+1- For more information about the codon usage bias (CUB), equations used to analyze CUB in the BCAW tool and the API `BCAW tool’s documentation <https://bcaw-tools-documentation.readthedocs.io/en/latest/index.html>`_.
+
 2. For more information about the output plots >> [Plots description](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Plots%20Description.md).
 
 3. For more information about the abbreviations used >> [Abbreviations table](https://github.com/AliYoussef96/BCAW-Tool/blob/master/Abbreviations.md)
