### 2019-07-23 eb60500
Update .travis.yml
