+++ .travis.yml
@@ -4,6 +4,5 @@ python:
   - "3.7"
 install:
   - pip install -r requirements.txt
-  - pip install -e .
 script:
   - pytest tests/test.py
