+++ paper/paper.md
@@ -28,7 +28,7 @@ BCAW tool was developed using python 3.7 with built in and third-party packages 
 
 ```
 from BCAWT import BCAWT
-BCAWT.BCAW('Ecoli','result_folder',genetic_code_=11,fasta=True,Auto=True)
+BCAWT.BCAW(['Ecoli.fasta'],'result_folder',genetic_code_=11,Auto=True)
 ```
 
 The expected outputs from the BCAW tool could be divided into three groups. The first one is data in CSV format described in (Table 1), the second group is plots summarized in (Fig 1), and the last one is text files each contains a different result for a different statistical test. The equations used for analyzing CUB in the tool, and the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
