### 2019-07-25 9be0b69
Update Dependencies.rst
