### 2019-07-25 09c8e4d
Update help.rst
