### 2019-07-25 d6d9576
Update help.rst
