+++ docs/help.rst
@@ -1,3 +1,3 @@
 Help
 ====
-If you need more help please contact `ali.mo.anwar@std.agr.cu.edu.eg <mailto:ali.mo.anwar@std.agr.cu.edu.eg`_.
+If you need more help please contact `ali.mo.anwar@std.agr.cu.edu.eg <mailto:ali.mo.anwar@std.agr.cu.edu.eg>`_.
