### 2019-07-25 dceed65
Update plots.rst
