### 2019-07-25 9101495
Update plots.rst
