### 2019-07-25 95d5bfc
Remove bold titles from docs and clarify language (#11)
* Update Contribution.rst

* Update Dependencies.rst

* Update Documentations.rst

* Update License.rst

* Update Statement.rst

* Update Table_output.rst

* Update Usage.rst

* Update Usage.rst

* Update help.rst

* Update help.rst

* Update install.rst

* Update intro.rst

* Update plots.rst

* Update plots.rst
