+++ docs/Table_output.rst
@@ -1,49 +1,49 @@
-**Outputs**
-===============
+Outputs
+=======
 
 
-**The expected CSV files output**
-------------------------------------
+Expected CSV output
+-------------------
 
 
 +------------------------------------+---------------------------------+
 | CSV file name                      | Description                     |
 +====================================+=================================+
-| ATCG                               | contains ; gene id, GC, GC1,    |
+| ATCG                               | Contains gene id, GC, GC1,      |
 |                                    | GC2, GC3, GC12, AT, AT3 A3, T3, |
-|                                    | C3, G3, GRAVY, AROMO and, Gene  |
-|                                    | Length                          |
+|                                    | C3, G3, GRAVY, AROMO and gene   |
+|                                    | length                          |
 +------------------------------------+---------------------------------+
-| CA_RSCU                            | contains ; each RSCU result for |
+| CA_RSCU                            | Each RSCU result for            |
 |                                    | each codon in each genes        |
 +------------------------------------+---------------------------------+
-| CA_RSCUcodons                      | contains ; correspondence       |
-|                                    | analysis first 4 axis for each  |
-|                                    | codon                           |
+| CA_RSCUcodons                      | Correspondence                  |
+|                                    | analysis for the first 4 axes of|
+|                                    | each codon                      |
 +------------------------------------+---------------------------------+
-| CA_RSCUgenes                       | contains ; correspondence       |
-|                                    | analysis first 4 axis for each  |
+| CA_RSCUgenes                       | Correspondence                  |
+|                                    | analysis for the first 4 axis of|
 |                                    | gene                            |
 +------------------------------------+---------------------------------+
-| CAI                                | contains ; gene id and CAI      |
-|                                    | index                           |
+| CAI                                | Gene id and CAI for each gene   |
+|                                    |                                 |
 +------------------------------------+---------------------------------+
-| ENc                                | contains ; gene id and ENc      |
-|                                    | index.                          |
+| ENc                                | Gene id and ENc for each gene   |
+|                                    |                                 |
 +------------------------------------+---------------------------------+
-| P2-index                           | contains ; gene id and P2 index |
+| P2-index                           | Gene id and P2 index for each   |
+|                                    | gene                            |
 +------------------------------------+---------------------------------+
-| optimal codons                     | contains; putative optimal      |
-|                                    | codons detected                 |
+| optimal codons                     | Putative optimal codons         |
 +------------------------------------+---------------------------------+
 
 
-**All output plots**
--------------------------
+All output plots
+----------------
 
 |image0|
 
-.. |image0| image:: https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/All%20plots.jpg
+.. |image0| image:: https://raw.githubusercontent.com/AliYoussef96/BCAW-Tool/master/Plots/All%20plots.jpg
    :width: 6.5in
    :height: 4.875in
    :alt: All output plots from BCAW tool analysis for coding sequence from Escherichia coli.
