+++ docs/plots.rst
@@ -1,4 +1,4 @@
-**More About plots**
-===========
+More About Plots
+================
 
-All output plots are described here >> https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Plots%20Description.md
+All output plots are described `here <https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Plots%20Description.md>`_
