+++ docs/Documentations.rst
@@ -1,7 +1,7 @@
-**More information**
-=======================
+More information
+================
 
-1. For more information about the codon usage bias (CUB), equations used to analyze CUB in the BCAW tool and the API `BCAW tool’s documentation <https://bcaw-tools-documentation.readthedocs.io/en/latest/index.html>`_.
+1. For more information about the codon usage bias (CUB) equations used to analyze CUB in the BCAW tool and the API `BCAW tool’s documentation <https://bcaw-tools-documentation.readthedocs.io/en/latest/index.html>`_.
 
 2. For more information about the output `plots <https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/Plots%20Description.md>`_.
 
