### 2019-07-25 37e586c
Update setup.py
