### 2019-08-02 6de3457
Update intro.rst
