+++ docs/intro.rst
@@ -149,87 +149,38 @@ affecting the synonymous codon usage within the genes group.
 
 References
 ----------
-1. Gu W, Zhou T, Ma J, Sun X, Lu Z. Analysis of synonymous codon usage
-in SARS Coronavirus and other viruses in the Nidovirales. Virus Res.
-2004;101: 155–161. doi:10.1016/j.virusres.2004.01.0062. Vicario S,
-Moriyama EN, Powell JR. Codon usage in twelve species of Drosophila. BMC
-Evol Biol. 2007;7: 1–17. doi:10.1186/1471-2148-7-2263. Behura SK,
-Severson DW. Comparative analysis of Codon usage bias and Codon context
-patterns between dipteran and hymenopteran sequenced genomes. PLoS One.
-2012;7. doi:10.1371/journal.pone.00431114. Boël G, Letso R, Neely H,
-Price WN, Su M, Luff J, et al. Codon influence on protein expression in
-E.coli. 2016;529: 358–363. doi:10.1038/nature16509.Codon5. Dohra H,
-Fujishima M, Suzuki H. Analysis of amino acid and codon usage in
-Paramecium bursaria. FEBS Lett. Federation of European Biochemical
-Societies; 2015;589: 3113–3118. doi:10.1016/j.febslet.2015.08.0336. Qiu
-S, Zeng K, Slotte T, Wright S, Charlesworth D. Reduced efficacy of
-natural selection on codon usage bias in selfing Arabidopsis and
-Capsella species. Genome Biol Evol. 2011;3: 868–880.
-doi:10.1093/gbe/evr0857. Wang L, Xing H, Yuan Y, Wang X, Saeed M, Tao J,
-et al. Genome-wide analysis of codon usage bias in four sequenced cotton
-species. PLoS One. 2018; 1–17. doi:10.1371/journal.pone.01943728. Chen
-H, Sun S, Norenburg JL, Sundberg P. Mutation and selection cause codon
-usage and bias in mitochondrial genomes of ribbon worms (Nemertea). PLoS
-One. 2014;9. doi:10.1371/journal.pone.00856319. Zalucki YM, Power PM,
-Jennings MP. Selection for efficient translation initiation biases codon
-usage at second amino acid position in secretory proteins. Nucleic Acids
-Res. 2007;35: 5748–5754. doi:10.1093/nar/gkm57710. Prabha R, Singh DP,
-Sinha S, Ahmad K, Rai A. Genome-wide comparative analysis of codon usage
-bias and codon context patterns among cyanobacterial genomes. Mar
-Genomics. Elsevier B.V.; 2017;32: 31–39.
-doi:10.1016/j.margen.2016.10.00111. Palidwor GA, Perkins TJ, Xia X. A
-general model of Codon bias due to GC mutational bias. PLoS One. 2010;5.
-doi:10.1371/journal.pone.001343112. Marais G, Mouchiroud D, Duret L.
-Neutral effect of recombination on base composition in Drosophila. Genet
-Res. 2003;81: 79–87. doi:10.1017/S001667230200607913. Rocha EPC. Codon
-usage bias from tRNA’s point of view: Redundancy, specialization, and
-efficient decoding for translation optimization. Genome Res. 2004;
-2279–2286. doi:10.1101/gr.289690414. Huang Y, Koonin E V., Lipman DJ,
-Przytycka TM. Selection for minimization of translational frameshifting
-errors as a factor in the evolution of codon usage. Nucleic Acids Res.
-2009;37: 6799–6810. doi:10.1093/nar/gkp71215. Duret L, Mouchiroud D.
[…]
