### 2019-08-04 #1500 @mgymrek
Thanks for the updates. @Benjamin-Lee can you also finish going throug the checklist after these revisions to see if remaining requirements are met?

One point: the pip version is 0.2.0 but the version listed on this review is v1.0.0. Which one is the most recent release?
