### 2019-08-05 #1500 @AliYoussef96
I updated the pip version to 1.0.0. to give the match with the GitHub release (v1.0.0).

[*The pip version*](https://pypi.org/project/BCAWT/)
