### 2019-08-05 #1500 @Benjamin-Lee
I’ll get on it!

> On Aug 5, 2019, at 10:04 AM, Ali Youssef <notifications@github.com> wrote:
>
> I updated the pip version to 1.0.0. to give the match with the GitHub release (v1.0.0).
>
> The pip version
>
