### 2019-08-15 7e5024d
Update intro.rst
