+++ docs/Statement.rst
@@ -1,9 +1,7 @@
 Statement of Need
 ==================
-There are no tools available enable users to run a whole automated workflow for codon usage bias analysis. 
 
-BCAW Tool (Bio Codon Analysis Workflow Tool) was developed to address this problem. 
+There are no tools available enable users to run a whole automated workflow for codon usage bias analysis. Using python 3.7 BCAW Tool ( Bio Codon Analysis Workflow Tool ) was developed to address this problem.
+BCAW Tool manages a complete automated workflow to analyze the codon usage bias for genes and genomes of any organism. With minimum coding skills.
 
-Expected results from BCAW are: nucleotide content, various codon usage measures (effective number of codons, codon adaptation index, *etc*.) as well as correlation analysis with graphs. In addition, correspondence analysis and determination of putative optimal codons are included. BCAW Tool manages a complete automated workflow to analyze the codon usage bias for the genes and genomes of any organism. BCAW Tool is available as executable application, works with the Windows operating system, and has its source code availible under the MIT License.
-
-For more details about CUB, and the equations used in BCAWT https://github.com/AliYoussef96/BCAW-Tool/blob/master/Introduction%20to%20codon%20usage%20bias.pdf.
+For more details about  codon usage bias , and the equations used in BCAWT see:https://bcaw-tools-documentation.readthedocs.io/en/latest/intro.html.
