### 2019-08-15 35ecb1b
Update statement.rst
