+++ docs/statement.rst
@@ -1,7 +1,6 @@
 Statement of Need
 ==================
 
-There are no tools available enable users to run a whole automated workflow for codon usage bias analysis. Using python 3.7 BCAW Tool ( Bio Codon Analysis Workflow Tool ) was developed to address this problem.
-BCAW Tool manages a complete automated workflow to analyze the codon usage bias for genes and genomes of any organism. With minimum coding skills.
+There are no tools available enable users to run a whole automated workflow for codon usage bias analysis. Using python 3.7 BCAW Tool ( Bio Codon Analysis Workflow Tool ) was developed to address this problem. BCAW Tool manages a complete automated workflow to analyze the codon usage bias for genes and genomes of any organism. With minimum coding skills.
 
-For more details about  codon usage bias , and the equations used in BCAWT see:https://bcaw-tools-documentation.readthedocs.io/en/latest/intro.html.
+For more details about CUB, and the equations used in BCAWT https://bcaw-tools-documentation.readthedocs.io/en/latest/intro.html
