### 2019-08-18 #1500 @AliYoussef96
@Benjamin-Lee and @mgymrek

About the checklist, any comments about the Functionality and the References?
