### 2019-08-18 #1500 @Benjamin-Lee
Not yet. Please be aware that I am on vacation starting out of internet service, so I will get back to you in a week!

> On Aug 18, 2019, at 1:19 PM, Ali Youssef <notifications@github.com> wrote:
>
> @Benjamin-Lee and @mgymrek
>
> About the checklist, any comments about the Functionality and the References?
>
