### 2019-09-12 #1500 @Benjamin-Lee
Hi! I’ll get on this tomorrow. Sorry for the delay.

> On Sep 12, 2019, at 1:12 AM, Melissa Gymrek <notifications@github.com> wrote:
>
> @AliYoussef96 I apologize for the delay in moving this forward. I have been waiting to hear back from the reviewer. Regardless of that I will update you with a decision by the end of this week.
>
