### 2019-09-13 AliYoussef96/BCAW-Tool#12 @Benjamin-Lee
Fails on Mac
I think it has to do with the fact that you're checking if the user is on Windows or Linux, which results in the unbound variable when on Mac.

```python
~ ❯❯❯ python test.py                                                                                                                    (bcaw)
Traceback (most recent call last):
  File "test.py", line 2, in <module>
    BCAWT.BCAW(['Ecoli.fasta'],'result_folder',genetic_code_=11,Auto=True)
  File "/usr/local/Caskroom/miniconda/base/envs/bcaw/lib/python3.7/site-packages/BCAWT/BCAWT.py", line 106, in BCAW
    if os.path.exists(directory) == False:
UnboundLocalError: local variable 'directory' referenced before assignment
~ ❯❯❯ ipython                                                                                                                           (bcaw)
Python 3.7.3 (default, Apr 13 2019, 18:43:32)
Type 'copyright', 'credits' or 'license' for more information
IPython 7.6.1 -- An enhanced Interactive Python. Type '?' for help.

In [1]: import platform

In [2]: platform.system()
Out[2]: 'Darwin'
```
