+++ BCAWT/BCAWT_auto_test.py
@@ -1,37 +1,48 @@
-def auto_test():
+def auto_test(path=str()):
     '''
+
     Run a demo test. creates a folder named "test_demo" with 23 results ( see the documentation for more details about the expected output )
+
+    Args:
+        path: absolute path to the directory to save the result in
 
     Returns:
         None
+
     '''
 
+    if path == str():
+        raise TypeError ("Must specify a path to save result in")
+
     from BCAWT import BCAWT
-    file =  open("Ecoli.fasta", "r")
+    file  =  open("Ecoli.fasta", "r")
 
-    BCAWT.BCAW(main_fasta_file = [file] , save_folder_name = 'test_demo',Auto=True)
+    BCAWT.BCAW(main_fasta_file = [file] , save_path= path, Auto=True)
 
 
 
-def auto_check_files():
+
+def auto_check_files(path):
     '''
     Check the expected outputs.
 
+    Args:
+        path: absolute path to the directory to save the result in
+
     Returns:
+
         None
     '''
-    import platform
-    import glob
-    if platform.system() == "Windows":
-        text = [ i for i in glob.glob("Result\\test_demo\\*.txt")]
-        csv = [ i for i in glob.glob("Result\\test_demo\\*.csv")]
-        png = [i for i in glob.glob("Result\\test_demo\\*.png")]
-    elif platform.system() == "Linux":
-        text = [ i for i in glob.glob("Result/test_demo/*.txt")]
-        csv = [ i for i in glob.glob("Result/test_demo/*.csv")]
-        png = [i for i in glob.glob("Result/test_demo/*.png")]
+    
+    import glob  
+    
+    text = [ i for i in glob.glob(f"{path}*.txt")]
+
+    csv = [ i for i in glob.glob(f"{path}*.csv")]
+
+    png = [i for i in glob.glob(f"{path}*.png")]
 
     if len(text) == 4 and len(csv) == 8 and len(png) == 11:
-        print ("test is completed 'successfully'")
+        print ("Test is completed 'successfully'")
     else:
         print ("Something going wrong please, see the documentation or contact the developer")
