+++ BCAWT/GC123.py
@@ -62,7 +62,7 @@ def GC12(sequ):
     GC2 = 0
 
     sequ = str(sequ)
-    codon, codon_1 = tee(sequ[i: i + 3] for i in range(0, len(sequ), 3))
+    codon, codon_1 = tee(sequ[i: i + 3] for i in range(0, len(sequ), 3) if len(sequ[i: i + 3]) == 3)
 
     lenght_codon = sum(1 for _ in codon_1)
 
