+++ BCAWT/BCAWT_auto_test.py
@@ -1,7 +1,7 @@
 def auto_test(path=str()):
     '''
 
-    Run a demo test. creates a folder named "test_demo" with 23 results ( see the documentation for more details about the expected output )
+    Run a demo test with 23 results ( see the documentation for more details about the expected output )
 
     Args:
         path: absolute path to the directory to save the result in
@@ -33,16 +33,17 @@ def auto_check_files(path):
 
         None
     '''
-    
+
     import glob  
     
-    text = [ i for i in glob.glob(f"{path}*.txt")]
+    text = [ i for i in glob.glob(f"{path}/*.txt")]
 
-    csv = [ i for i in glob.glob(f"{path}*.csv")]
+    csv = [ i for i in glob.glob(f"{path}/*.csv")]
 
-    png = [i for i in glob.glob(f"{path}*.png")]
+    png = [i for i in glob.glob(f"{path}/*.png")]
 
     if len(text) == 4 and len(csv) == 8 and len(png) == 11:
+        
         print ("Test is completed 'successfully'")
     else:
         print ("Something going wrong please, see the documentation or contact the developer")
