### 2019-09-15 8ae3610
Update BCAWT.py
