### 2019-09-15 AliYoussef96/BCAW-Tool#12 @AliYoussef96
I replaced one parameter in the main function, I believe this will solve the issue and will be compatible with any platform.

The new parameter (save_path) will take an absolute path to a directory to save the result in, with default value equal to a current directory.

https://github.com/AliYoussef96/BCAW-Tool/blob/9c03019bfcdeeeb783aa3f2336c163ec30e0c42d/BCAWT/BCAWT.py#L1

I did test the code on Windows10 and Ubuntu19 and it works fine. I don't have access to a Mac right now but with the current changes, I think It will work fine with you.

I updated the documentation, README, as well as the pip version.
**pip version now is 1.0.1.**
