### 2019-09-15 #1500 @AliYoussef96
> Just worked on trying it out and am currently running into an error on Mac. Once I can review the functionality, we'll be good to go.

Hi @Benjamin-Lee

I made changes in the main function and I believe this will solve the error.

https://github.com/AliYoussef96/BCAW-Tool/issues/12#issuecomment-531560160
