### 2019-09-20 #1500 @mgymrek
Hello,
I am testing out these changes. Using the version from pip and the example:

```from BCAWT import BCAWT_auto_test
path = "Test_folder" # or any!!
BCAWT_auto_test.auto_test(path)
#processing....
BCAWT_auto_test.auto_check_files(path)
```

I get an error:
```
Bio.Data.CodonTable.TranslationError: Codon '<DI' is invalid
```

The main example (`BCAWT.BCAW(['Ecoli.fasta'],'save_path',genetic_code_=11,Auto=True)`) worked smoothly.
