### 2019-09-20 #1500 @AliYoussef96
Hello @mgymrek,

The error you got from biopython translate function when it tries to translate unknown codons.  Any invalid codon (e.g. "TA?" or "T-A") will throw a TranslationError.
link to the error description : https://biopython.org/DIST/docs/api/Bio.Seq-module.html#translate

According to the error, a codon "**<DI**" is in one of DNA sequences in the fasta file being tested, I'm sure the [test file](https://github.com/AliYoussef96/BCAW-Tool/blob/master/tests/Ecoli.fasta) I added on the repository has valid DNA sequences.

The automatic testing you trying to run expect this file.
```python
from BCAWT import BCAWT_auto_test
path = "Test_folder" # or any!!
BCAWT_auto_test.auto_test(path)
#processing....
BCAWT_auto_test.auto_check_files(path)
```
So, kindly can you make sure you have the write file.
