### 2019-09-22 #1500 @mgymrek
Ok, I got it to work for now. The usage on the site is misleading. I assumed "Test_folder" is a folder containing the fasta files to check. (The functionality seemed to change since I tried a couple days ago using the pip version, since earlier I gave a directory and it didn't complain that it couldn't find the Ecoli.fasta file)
The README mentions any fasta file can be downloaded under "Auto testing". However "Ecoli.fasta" appears to be hardcoded into the test function. The usage should clarify whether that specific file is needed. If it is, that should be part of the default install so that users don't have to seprately fetch that file to test.

Also, here are some comments on the paper:
"have been developed to analysis" -> "have been developed to analyze"
"of any organism genes" -> "of any organism's genes"
"Further, Using" -> "Further, using"
"CUB as;" -> "CUB such as;"
"statistical analysis as" -> "statistical analysis such as"
"with build in" -> "with built in"
"a bunch of analysis" -> "a variety of analyses"

Once the above are taken care of we should be ready to go.
