### 2019-09-22 24bf6d6
Update paper.md
