### 2019-09-22 4d6e6aa
Update BCAWT_auto_test.py
