### 2019-09-22 #1500 @AliYoussef96
1- I updated your comments on the paper.

2- I changed the test function to accept any fasta file and I made it clear in the README.

```python
from BCAWT import BCAWT_auto_test
path = "Test_folder" # absolute path to the directory to save the result in
test_fasta = "Test_fasta_file" # absolute path to the fasta file that will be tested
BCAWT_auto_test.auto_test(path, test_fasta)
#processing....
BCAWT_auto_test.auto_check_files(path) # note: this test assumes that in the result folder nothing except the result files form the above function.
```
3- The pip version now it 1.0.3.
