### 2019-09-24 #1500 @danielskatz
You might try changing the path to the image in the .md file to

`https://github.com/AliYoussef96/BCAW-Tool/raw/master/Plots/All%20plots.jpg`

the current link is to https://github.com/AliYoussef96/BCAW-Tool/blob/master/Plots/All%20plots.jpg which is the github page containing the image, not the raw image itself
