### 2019-09-24 #1500 @AliYoussef96
@mgymrek I changed the path to the image in the .md, and it worked.
