### 2019-09-27 #1500 @AliYoussef96
@mgymrek Can I ask what is next? are we ready to go further?
