### 2019-10-01 #1500 @mgymrek
Thanks for fixing the proof. We are almost there. Last step: Can you now create a Zenodo archive and report the DOI in this thread. Please make sure the Zenodo archive has the same title and author list as the paper. Also please indicate whether the version needs to be updated (it is listed as v1.0.0 on this thread).
