### 2019-10-01 #1500 @AliYoussef96
Hi @mgymrek, thank you for your response.

I created the Zenodo archive, DOI:  [![DOI](https://zenodo.org/badge/189912439.svg)](https://zenodo.org/badge/latestdoi/189912439) . With the same title and author list as the [paper](https://github.com/openjournals/joss-papers/blob/joss.01500/joss.01500/10.21105.joss.01500.pdf).

Yes, I want to update the version to v1.0.3 as pip [version](https://pypi.org/project/BCAWT/1.0.3/], I tried to do it, but I see only the editors are allowed to do that.
