+++ paper/paper.md
@@ -18,13 +18,13 @@ bibliography: paper.bib
 
 The redundancy in the genetic code means that apart from methionine and tryptophan, an amino acid is encoded by at least two codons. Different codons for the same amino acid are termed synonymous codons. Synonymous codon usage is strongly influenced by evolutionary forces namely, selection and mutation and may vary strongly within or among organisms The preference of specific codons over others contributes to this variation and this phenomenon is called codon usage bias (CUB).
 
-Many measurements have been developed to analyze and study CUB; effective number of codons ( ENc ) [@Wright1990], codon adaptation index ( CAI ), relative synonymous codon usage ( RSCU ) [@Sharp1987] and, translational selection index ( P2-index ) [@Wang2018]. Also, statistical analysis has been used to investigate the effect of different factors as selection and mutation on shaping CUB such as; Correspondence analysis, Parity Rule 2-plot Analysis and, Neutrality Plot [@Song2017]. BCAW tool was developed to analyze such phenomena ( Codon Usage Bias ) by the aforementioned  measurements.
+Many measurements have been developed to analyze and study CUB; effective number of codons (ENc) [@Wright1990], codon adaptation index (CAI), relative synonymous codon usage (RSCU) [@Sharp1987] and, translational selection index (P2-index) [@Wang2018]. Also, statistical analysis has been used to investigate the effect of different factors as selection and mutation on shaping CUB such as; Correspondence analysis, Parity Rule 2-plot Analysis and, Neutrality Plot [@Song2017]. BCAWT was developed to analyze such phenomena (Codon Usage Bias) by the aforementioned  measurements.
 
-Various tools are available to analyze and measure CUB, but they lack some important measurements and plots for CUB analysis. What BCAW tool does is an automated workflow to study the CUB of any organism's genes by all the measurements and plots mentioned above. Further, using the correlation method to determine the optimal codons described by [@Hershberg2009] is implemented for the first time in BCAW tool. The tool also includes statistical analysis such as correspondence analysis, correlation analysis, and t-test.
+Various tools are available to analyze and measure CUB, but they lack some important measurements and plots for CUB analysis. What BCAWT does is an automated workflow to study the CUB of any organism's genes by all the measurements and plots mentioned above. Furthermore, using the correlation method to determine the optimal codons described by @Hershberg2009 is implemented for the first time in BCAWT. The tool also includes statistical analysis such as correspondence analysis, correlation analysis, and t-test.
  
 # Implementation
 
-BCAW tool was developed using python 3.7 with built in and third-party packages [@Lee2018]. The usage of BCAW Tool is made to be very easy where users have only to input a fasta format file containing genes to be analyzed, and variety of analyses will be performed.
+BCAWT was developed using python 3.7 with built-in and third-party packages [@Lee2018]. The usage of BCAWT is made to be very easy where users have only to input a fasta format file containing genes to be analyzed, and variety of analyses will be performed.
 
 ```python
 from BCAWT import BCAWT
@@ -32,15 +32,15 @@ BCAWT.BCAW(['Ecoli.fasta'],'save_path',genetic_code_=11,Auto=True)
 # processing...
 ```
 
-The expected outputs from the BCAW tool could be divided into three groups. The first one is data in CSV format described in (Table 1), the second group is plots summarized in (Fig 1), and the last one is text files each contains a different result for a different statistical test. The equations used for analyzing CUB in the tool, and the API are reported in the BCAWT's [documentation](https://bcaw-tools-documentation.readthedocs.io/en/latest/).
[…]
