### 2019-10-18 b10a9c9
Update paper.md
