+++ paper/paper.md
@@ -58,6 +58,6 @@ Furthermore, BCAWT returns 11 plots (Fig 1), enabling an easy interpretation of 
 
 **Fig 1: All output plots from BCAWT analysis for coding sequence from Escherichia coli.**
 
-![Fig 1](https://raw.githubusercontent.com/AliYoussef96/BCAW-Tool/master/Plots/All%20plots.jpg)
+![Fig 1](https://raw.githubusercontent.com/AliYoussef96/BCAW-Tool/master/Plots/all.png)
 
 # References
