### 2019-10-19 #1500 @AliYoussef96
- [x] Can you remove the extra spaces introduced for items in brackets throughout the paper?
- [x] Please amend your citation syntax for ...codons described by (Hersh-berg & Petrov, 2009)
- [x] Add hyphen to built in in: ...with built in and third-party packages...
- [x] Use can instead of could in "...tool could be divided into three groups."
- [x] Should BCAWT’s be BCAW’s? in "...are reported in the BCAWT’s documentation."
- [x] Consider rephrasing this sentence, it does not read well and may be confusing "The first one is data in CSV format described in (Table 1), the second group is plots summarized in (Fig 1), and the last one is text files each contains a different result for a different statistical test.". I propose (but correct it if I misunderstood): "The first group is data in the CSV format (see Table 1), the second group is plots (summarized in Fig 1), and the last group is text files, whereby each text file contains results for a different statistical test."
- [x] Using 1-, 2- for a list in a sentence is unusual. Please change to: 1), 2).
- [x] Please rephrase this sentence which does not read well: The advantages for BCAW tool over the existing tools are; 1- Automated Workflow, 2- No limitations on the number of genes, 3- The method used to determine optimal codons named correlation method is only available in the BCAW tool, 4- Plots as violin plots for nucleotides contents, and other plots are generated by the BCAW tool and no need for using another software to draw such plots. I propose (but correct it if I misunderstood): The advantages of BCAW tool over existing tools are; 1) the automated workflow, 2) the ability to handle large numbers of genes, 3) the method used to determine optimal codons, named the correlation method, is only available in the BCAW tool, 4) visualization and plotting capability, including the creation of violin plots for nucleotide contents, removing the need for other plotting software.
- [x] Can you increase the fonts used in figure 1?
- [x] Instead of Further, can you sue Furthermore, in "Further, using the correlation method..."
- [x] You use both the BCAW tool and BCAW tool. Both read more awkward then BCAWT. I propose you use BCAWT throughout the paper instead of the aforementioned.
- [x] Add a in BCAW tool returns CSV file containing... Such that it reads: BCAW tool returns a CSV file containing...
- [x] In ....containing the codon usage bias indices output ( Table 1 )., it seems you can use the acronym CUB instead.
- [x] In the table title: Table 1: Explain the CSV files output..., rephrase this to Table 1: Explanation of the the CSV output file...
- [x] Please rephrase Also, it returns 11 plot ( Fig 1 ), so one could interpret the results easily. to Furthermore, BCAWT returns 11 plots (Fig 1), enabling easy interpretation of the results.
- [x] Can you check the title for figure 1 Fig 1: All output plots from BCAW tool analysis for coding sequence from Escherichia, should that be "codon sequence"? or "a code sequence"? or "genetic code sequence"?  >>> I meant ``` genes coding sequence```
- [x] Please rephrase The usage of BCAW Tool is made to be very easy where users have only to input a fasta format file containing genes to be analyzed, and variety of analyses will be performed.
- [x] Make sure you have an enter and spaces after the figure label to ensure the figure does not back the figure label (as it does not)
