### 2019-10-19 #1500 @AliYoussef96
@Kevin-Mattheus-Moerman Thank you for your quick response!!

- [x] Please use either the BCAWT throughout or BCAWT. I propose BCAWT, so to remove all instances of the in front of BCAWT.
- [x] Check again for BCAW and BCAWT (e.g. you still have ....are reported in the BCAW’s)
- [x] You currently do not have an acknowledgements section (which is optional), add one now if you feel it is appropriate. >>> It is fine without.
- [x] You currently have two figure labels, i.e. above and below the figure. Can you use just one label under the figure?
- [x] Can you rephrase: For example, to analyze genes coding sequence within Ecoli.fasta, save the results to a folder named save_path and choosing genetic code for bacteria:, I propose (unless I misunderstood): For example, the following code snippet shows how to analyze genes for a coding sequence within the file `Ecoli.fasta`, for a genetic code specified, and to save the results to a folder named save_path.
