### 2019-10-20 #1500 @Kevin-Mattheus-Moerman
@AliYoussef96 getting close
- [ ] Add a full stop after `... a folder named save_path` so it reads `... a folder named save_path.`
- [ ] Remove the `the` in `...are reported in the BCAW’s` so it becomes: `.....are reported in BCAW’s`
