### 2019-10-20 #1500 @AliYoussef96
• [x] Add a full stop after ... a folder named save_path so it reads ... a folder named save_path.
• [x] Remove the the in ...are reported in the BCAW’s so it becomes: .....are reported in BCAW’s
